chore(folder-structure) setup folder structure for Authors Haven project - Create new sar- Create a `server` folder for organizing server codebase - Add the relevant folders to `server` folder - Add the corresponding files and code - Update the import references - Refactor codebase to separate controllers from routes [Delivers #159316769]
