+++ .eslintignore
@@ -1,8 +1,8 @@
 /node_modules/**
 /coverage/**
-/migrations/**
-/seeders/**
-/test/**
+**/migrations/**
+**/seeders/**
+**/test/**
 .env
 package-lock.json
