+++ .sequelizerc
@@ -1,8 +1,8 @@
 const path = require('path');
 
 module.exports = {
-  'config': path.resolve('config', 'config.js'),
-  'models-path': path.resolve('./models'),
-  'seeders-path': path.resolve('./seeders'),
-  'migrations-path': path.resolve('./migrations')
+  'config': path.resolve('./server/configs', 'config.js'),
+  'models-path': path.resolve('./server/database/models'),
+  'seeders-path': path.resolve('./server/database/seeders'),
+  'migrations-path': path.resolve('./server/database/migrations')
 };
