--- config/index.js
@@ -1,5 +1,0 @@
-
-export default {
-  secret:
-        process.env.NODE_ENV === 'production' ? process.env.SECRET : 'secret',
-};
