+++ server/database/seeders/20180731151422-demo-user.js
@@ -1,4 +1,4 @@
-
+/* eslint-disable */
 
 module.exports = {
   up: (queryInterface, Sequelize) => queryInterface.bulkInsert('Users', [{
@@ -6,9 +6,8 @@ module.exports = {
     email: 'johnDoe@mail.com',
     hash: '12345678',
     createdAt: new Date(),
-    updatedAt: new Date()
+    updatedAt: new Date(),
   }], {}),
-  // eslint-disable-next-line  no-unused-vars
   down:
-  (queryInterface, Sequelize) => queryInterface.bulkDelete('Users', null, {})
+  (queryInterface, Sequelize) => queryInterface.bulkDelete('Users', null, {}),
 };
