+++ server/test/routes/index.test.js
@@ -1,6 +1,9 @@
+/* eslint-disable */
+
 import chai, { expect, use, request } from 'chai';
 import chaiHttp from 'chai-http';
-import server  from '../index';
+import server from '../../index';
+
 use(chaiHttp);
 
 describe('Testing the index.js file', () => {
