+++ server/test/routes/mockData.js
@@ -1,4 +1,4 @@
- 
+/* eslint-disable */
 
 const mockData = {
   user1: {
